重构thumbnail_loader
